Agent: Move ExploiterResultData to its own file under i_puppet
